% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createCohorts}
\alias{createCohorts}
\title{Create Target and Comparator Cohorts}
\usage{
createCohorts(cdmDb, scratchDatabaseSchema, connectionDetails)
}
\arguments{
\item{connectionDetails}{}
}
\description{
Create Target and Comparator Cohorts
}
